Use LibraryLoader in WindowsFFI, and correct GetProcessId() signature for win64

+++ src/org/jruby/util/ShellLauncher.java
@@ -630,7 +630,7 @@ public class ShellLauncher {
                                 return (Integer)UNIXProcess_pid.get(process);
                             } else if (ProcessImpl.isInstance(process)) {
                                 Long hproc = (Long) ProcessImpl_handle.get(process);
-                                return WindowsFFI.getKernel32().GetProcessId(hproc.intValue());
+                                return WindowsFFI.getKernel32().GetProcessId(hproc);
                             }
                         } catch (Exception e) {
                             // ignore and use hashcode
@@ -660,7 +660,7 @@ public class ShellLauncher {
                     try {
                         if (ProcessImpl.isInstance(process)) {
                             Long hproc = (Long) ProcessImpl_handle.get(process);
-                            return WindowsFFI.getKernel32().GetProcessId(hproc.intValue());
+                            return WindowsFFI.getKernel32().GetProcessId(hproc);
                         }
 
                     } catch (Exception e) {
